:truck: Move YouTube store to misc/deprecated
And also:
  - :truck: Move `youtubedl` module to misc/deprecated
  - :fire: remove related YouTube Store documentation (rst files)
  - :wrench: Adapt config files to the removal of Picasa
