--- docs/source/backends/youtube.rst
@@ -1,7 +1,0 @@
-
-.. _coherence.backends.youtube:
-
-coherence.backends.youtube\_storage (module)
---------------------------------------------
-
-.. automodule:: coherence.backends.youtube_storage
